Rename integrate app to bind
