--- integrate/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class IntegrateConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "integrate"
-    verbose_name = "Integration Models"
