+++ pyproject.toml
@@ -23,4 +23,4 @@ Repository = "https://github.com/arthexis/arthexis"
 Homepage = "https://arthexis.com"
 
 [tool.setuptools]
-packages = [ "core", "config", "nodes", "ocpp", "website", "release", "integrate",]
+packages = [ "core", "config", "nodes", "ocpp", "website", "release", "bind",]
