+++ release/utils.py
@@ -112,7 +112,7 @@ def _write_pyproject(package: Package, version: str, requirements: list[str]) ->
                     "ocpp",
                     "website",
                     "release",
-                    "integrate",
+                    "bind",
                 ]
             }
         },
